emeail_parser/requirements.txt
langchain==0.2.14
langgraph==0.0.23
pydantic==2.8.2
python-dotenv==1.0.1
email-validator==2.1.0.post1
pytest==8.2.1
rich==13.8.0
google-generativeai==0.7.2
langchain-google-genai==1.0.3